builds functional NameNode/DataNodes
this is still a broken state - the JobTracker and TaskTrackers
do not start. this is probably due to CDH4 rather than due to
script misconfiguration. the JobTracker doesn't start, but logs
nothing into /var/log/hadoop, so will be a little more subtle
to figure out - this will be enough to get HBase working.